appinsights/requirements.txt
config
azure-applicationinsights
opencensus-ext-azure